+++ jena-tdb/src/main/java/org/apache/jena/tdb/solver/stats/StatsResults.java
@@ -25,23 +25,23 @@ import org.apache.jena.graph.Node ;
 /** Statistics collector */
 public class StatsResults 
 {
-    private final Map<Node, Integer> predicates ;
-    private final Map<Node, Integer> types ;
+    private final Map<Node, Long> predicates ;
+    private final Map<Node, Long> types ;
     private final long count ;
 
-    StatsResults(Map<Node, Integer> predicates, Map<Node, Integer> types, long count)
+    StatsResults(Map<Node, Long> predicates, Map<Node, Long> types, long count)
     {
         this.count = count ;
         this.predicates = predicates ;
         this.types = types ;
     }
 
-    public Map<Node, Integer> getPredicates()
+    public Map<Node, Long> getPredicates()
     {
         return predicates ;
     }
 
-    public Map<Node, Integer> getTypes()
+    public Map<Node, Long> getTypes()
     {
         return types ;
     }
